+++ cli/cli/commands/enclave/inspect/user_services.go
@@ -3,67 +3,69 @@ package inspect
 import (
 	"context"
 	"fmt"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager/types"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/container_status"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/output_printers"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/forever_constants"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
+	"sort"
+	"strings"
 )
 
 const (
 	userServiceGUIDColHeader                                    = "GUID"
-	userServiceIDColHeader                                      = "ID"
-	userServiceHostMachinePortBindingsColHeader                 = "LocalPortBindings"
-	shouldShowStoppedContainersWhenGettingUserServiceContainers = true
+	userServiceIDColHeader    = "ID"
+	userServicePortsColHeader = "Ports"
 
-	noUserServiceHostPortBindingsPlaceholder = "<none>"
+	missingPortPlaceholder = "<none>"
 )
 
-func printUserServices(ctx context.Context, dockerManager *docker_manager.DockerManager, enclaveId string) error {
-	userServiceLabels := getLabelsForListEnclaveUserServices(enclaveId)
+func printUserServices(ctx context.Context, kurtosisBackend backend_interface.KurtosisBackend, enclaveId enclave.EnclaveID) error {
 
-	containers, err := dockerManager.GetContainersByLabels(ctx, userServiceLabels, shouldShowStoppedContainersWhenGettingUserServiceContainers)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting user service containers by labels: '%+v'", userServiceLabels)
+	userServiceFilters := &service.ServiceFilters{
+		EnclaveIDs: map[enclave.EnclaveID]bool{
+			enclaveId: true,
+		},
 	}
 
-	tablePrinter := output_printers.NewTablePrinter(userServiceGUIDColHeader, userServiceIDColHeader, userServiceHostMachinePortBindingsColHeader)
-	sortedContainers, err := sortContainersByGUID(containers)
+	userServices, err := kurtosisBackend.GetUserServices(ctx, userServiceFilters)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred sorting user service containers by GUID")
+		return stacktrace.Propagate(err, "An error occurred getting user services using filters '%+v'", userServiceFilters)
 	}
-	for _, container := range sortedContainers {
-		serviceGuid, found := container.GetLabels()[schema.GUIDLabel]
-		if !found {
-			return stacktrace.NewError("No '%v' container label was found in container ID '%v' with labels '%+v'", schema.GUIDLabel, container.GetId(), container.GetLabels())
-		}
-		serviceId, found := container.GetLabels()[schema.IDLabel]
-		if !found {
-			return stacktrace.NewError("No '%v' container label was found in container ID '%v' with labels '%+v'", schema.IDLabel, container.GetId(), container.GetLabels())
-		}
 
-		hostPortBindingsStrings := getContainerHostPortBindingStrings(container)
-		firstHostPortBindingStr := noUserServiceHostPortBindingsPlaceholder
-		if hostPortBindingsStrings != nil {
-			firstHostPortBindingStr = hostPortBindingsStrings[0]
-			hostPortBindingsStrings = hostPortBindingsStrings[1:]
+	tablePrinter := output_printers.NewTablePrinter(
+		userServiceGUIDColHeader,
+		userServiceIDColHeader,
+		userServicePortsColHeader,
+	)
+	sortedUserServices:= getSortedUserServiceSliceFromUserServiceMap(userServices)
+	for _, userService := range sortedUserServices {
+
+		guidStr := string(userService.GetGUID())
+		idStr := string(userService.GetID())
+		portBindingLines, err := getPortBindingStrings(userService)
+		if err != nil {
+			return stacktrace.Propagate(err, "An error occurred getting the port binding strings")
 		}
-		if err := tablePrinter.AddRow(serviceGuid, serviceId, firstHostPortBindingStr); err != nil {
+		firstPortBindingLine := portBindingLines[0]
+		additionalPortBindingLines := portBindingLines[1:]
+
+		if err := tablePrinter.AddRow(guidStr, idStr, firstPortBindingLine); err != nil {
 			return stacktrace.Propagate(
 				err,
-				"An error occurred adding row for user service container '%v' to the table printer",
-				serviceGuid,
+				"An error occurred adding row for user service with GUID '%v' to the table printer",
+				guidStr,
 			)
 		}
 
-		for _, additionalHostPortBindingStr := range hostPortBindingsStrings {
-			if err := tablePrinter.AddRow("", "", additionalHostPortBindingStr); err != nil {
+		for _, additionalPortBindingLine := range additionalPortBindingLines {
+			if err := tablePrinter.AddRow("", "", additionalPortBindingLine); err != nil {
 				return stacktrace.Propagate(
 					err,
-					"An error occurred adding additional host port binding '%v' row for user service container '%v' to the table printer",
-					additionalHostPortBindingStr,
-					serviceGuid,
+					"An error occurred adding additional port binding row '%v' for user service with GUID '%v' to the table printer",
+					additionalPortBindingLine,
+					guidStr,
 				)
 			}
 		}
@@ -73,19 +75,64 @@ func printUserServices(ctx context.Context, dockerManager *docker_manager.Docker
 	return nil
 }
 
-func getContainerHostPortBindingStrings(container *types.Container) []string {
-	var allHosPortBindings []string
-	hostPortBindings := container.GetHostPortBindings()
-	for hostPortBindingKey, hostPortBinding := range hostPortBindings {
-		hostPortBindingString := fmt.Sprintf("%v -> %v:%v", hostPortBindingKey, hostPortBinding.HostIP, hostPortBinding.HostPort)
-		allHosPortBindings = append(allHosPortBindings, hostPortBindingString)
+func getSortedUserServiceSliceFromUserServiceMap(userServices map[service.ServiceGUID]*service.Service) []*service.Service {
+	userServicesResult := make([]*service.Service, 0, len(userServices))
+	for _, userService := range userServices {
+		userServicesResult = append(userServicesResult, userService)
 	}
-	return allHosPortBindings
+
+	sort.Slice(userServicesResult, func(i, j int) bool {
+		return userServicesResult[i].GetGUID() < userServicesResult[j].GetGUID()
+	})
+
+	return userServicesResult
 }
 
-func getLabelsForListEnclaveUserServices(enclaveId string) map[string]string {
-	labels := map[string]string{}
-	labels[forever_constants.ContainerTypeLabel] = schema.ContainerTypeUserServiceContainer
-	labels[schema.EnclaveIDContainerLabel] = enclaveId
-	return labels
+// Guaranteed to have at least one entry
+func getPortBindingStrings(userService *service.Service) ([]string, error) {
+	privatePorts := userService.GetPrivatePorts()
+	if len(privatePorts) == 0 {
+		return []string{missingPortPlaceholder}, nil
+	}
+
+	portIds := []string{}
+	resultLines := map[string]string{}
+	for portId, privatePortSpec := range privatePorts {
+		portIds = append(portIds, portId)
+		line := fmt.Sprintf(
+			"%v: %v/%v",
+			portId,
+			privatePortSpec.GetNumber(),
+			strings.ToLower(privatePortSpec.GetProtocol().String()),
+		)
+		resultLines[portId] = line
+	}
+
+	// If the container is running, add host machine port binding information
+	if userService.GetStatus() == container_status.ContainerStatus_Running {
+		publicIpAddr := userService.GetMaybePublicIP()
+		publicPorts := userService.GetMaybePublicPorts()
+		for portId := range privatePorts {
+			publicPortSpec, found := publicPorts[portId]
+			if !found {
+				return nil, stacktrace.NewError(
+					"Private port '%v' was declared on service '%v' and the container is running, but no corresponding public port " +
+						"was found; this is very strange!",
+					portId,
+					userService.GetGUID(),
+				)
+			}
+			currentPortLine := resultLines[portId]
+			resultLines[portId] = currentPortLine + fmt.Sprintf(" -> %v:%v", publicIpAddr, publicPortSpec.GetNumber())
+		}
+	}
+
+	// Finally, sort the resulting lines by port ID
+	sort.Strings(portIds)
+	result := []string{}
+	for _, portId := range portIds {
+		result = append(result, resultLines[portId])
+	}
+
+	return result, nil
 }